move hooks to modules
