+++ assets/js/index.js
@@ -1,0 +1,2 @@
+export { default as Carousel } from "./hooks/carousel.js";
+export { default as Tabs } from "./hooks/tabs.js";
