+++ demo/assets/build.mjs
@@ -7,7 +7,6 @@ import { sassPlugin } from "esbuild-sass-plugin";
 const args = process.argv.slice(2);
 const watch = args.includes("--watch");
 const production = args.includes("--production");
-const mixEnv = production ? "prod" : "dev";
 
 const postCssPlugins = [autoprefixer];
 
@@ -30,9 +29,6 @@ let opts = {
   target: "es2017",
   outdir: "../priv/static/assets",
   plugins: plugins,
-  alias: {
-    "phoenix-colocated/doggo": `../_build/${mixEnv}/phoenix-colocated/doggo`,
-  },
 };
 
 if (production) {
